make custom hooks for PillView
