+++ src/renderer/src/views/hooks/useIdleOpacity.ts
